Fix DelayedInit behaviour when multiple racing gets exist
Previously only one get() would be selected as the "channel poller",
with all other futures effectively stalling until that one completed
(or was cancelled). This lead to suspended futures being able to deadlock
all other clients.

This fixes the issue by only holding the lock for the duration of the poll
call.